fixed (chapter 3) In exercises 1 fixed the average

+++ src/capitulo3/bloque03/Ejercicio01_SumaMediaYSalirseConCero.java
@@ -25,6 +25,6 @@ public class Ejercicio01_SumaMediaYSalirseConCero {
 		}
 		
 		System.out.println("La suma de sus números es: " + sum);
-		System.out.println("La media de sus números es: " + sum/(i-1));
+		System.out.println("La media de sus números es: " + sum/(float) (i-1));
 	}
 }
